feat: eslint.config.js 생성 \n\n 1. eslintrc.json 은 deprecated 되었으므로 eslint.config.js 로 작업 \n 2. eslint ignore file 들은 prepare-commit-msg hook에서 에러를 내뿜지 않도록 setting
